+++ buildSrc/src/main/kotlin/io/getstream/chat/android/Dependencies.kt
@@ -19,7 +19,7 @@ object Versions {
     internal const val ANDROIDX_TEST_JUNIT = "1.1.3"
     internal const val ANDROIDX_TEST = "1.4.0"
     internal const val ANDROIDX_VIEW_PAGER_2 = "1.0.0"
-    internal const val COIL = "1.4.0"
+    internal const val COIL = "2.1.0"
     internal const val COMPOSE_ACCOMPANIST = "0.23.0"
     internal const val COMPOSE_ACTIVITY = "1.4.0"
     internal const val COMPOSE_VIEW_MODEL = "2.4.1"